Commit message

+++ servos.py
@@ -137,18 +137,18 @@ class ServosThread(threading.Thread):
     
     def setRotate(self, rotate):
         self.position['rotate'] = rotate
-        self.servos[SERVO_BOARD][LEFT_ROTATE_PIN] = mapToLimits((self.position['rotate'] + LEFT_ROTATE_OFFSET)  * LEFT_ROTATE_DIRECTION)
-        self.servos[SERVO_BOARD][RIGHT_ROTATE_PIN] = mapToLimits((self.position['rotate'] + RIGHT_ROTATE_OFFSET) * RIGHT_ROTATE_DIRECTION)
+        self.servos[SERVO_BOARD][LEFT_ROTATE_PIN] = int(mapToLimits((self.position['rotate'] + LEFT_ROTATE_OFFSET)  * LEFT_ROTATE_DIRECTION))
+        self.servos[SERVO_BOARD][RIGHT_ROTATE_PIN] = int(mapToLimits((self.position['rotate'] + RIGHT_ROTATE_OFFSET) * RIGHT_ROTATE_DIRECTION))
     
     def setLift(self, lift):
         self.position['lift'] = lift
-        self.servos[SERVO_BOARD][LEFT_LIFT_PIN] = mapToLimits((self.position['lift'] + LEFT_LIFT_OFFSET)  * LEFT_LIFT_DIRECTION)
-        self.servos[SERVO_BOARD][RIGHT_LIFT_PIN] = mapToLimits((self.position['lift'] + RIGHT_LIFT_OFFSET) * RIGHT_LIFT_DIRECTION)
+        self.servos[SERVO_BOARD][LEFT_LIFT_PIN] = int(mapToLimits((self.position['lift'] + LEFT_LIFT_OFFSET)  * LEFT_LIFT_DIRECTION))
+        self.servos[SERVO_BOARD][RIGHT_LIFT_PIN] = int(mapToLimits((self.position['lift'] + RIGHT_LIFT_OFFSET) * RIGHT_LIFT_DIRECTION))
         
     def setGrab(self, grab):
         self.position['grab'] = grab
-        self.servos[SERVO_BOARD][LEFT_GRAB_PIN] = mapToLimits((self.position['grab'] + LEFT_GRAB_OFFSET)  * LEFT_GRAB_DIRECTION)
-        self.servos[SERVO_BOARD][RIGHT_GRAB_PIN] = mapToLimits((self.position['grab'] + RIGHT_GRAB_OFFSET) * RIGHT_GRAB_DIRECTION)
+        self.servos[SERVO_BOARD][LEFT_GRAB_PIN] = int(mapToLimits((self.position['grab'] + LEFT_GRAB_OFFSET)  * LEFT_GRAB_DIRECTION))
+        self.servos[SERVO_BOARD][RIGHT_GRAB_PIN] = int(mapToLimits((self.position['grab'] + RIGHT_GRAB_OFFSET) * RIGHT_GRAB_DIRECTION))
     
     def run(self):
         print "Starting " + self.name
